fix(#129): @TransactionalEventListener 전환 및 이벤트 패키지 통일
- @EventListener → @TransactionalEventListener(AFTER_COMMIT) 변경
- 알림 실패가 비즈니스 트랜잭션에 영향 미치지 않도록 분리
- 이벤트 패키지를 core.domain.event로 통일 (core.event → core.domain.event)
- 관련 import 전체 수정 (GameScorerServiceImpl, TeamMembershipServiceImpl, AttendanceServiceImpl, LineupService 및 각 테스트)
