+++ nextup-core/src/main/kotlin/com/nextup/core/domain/event/LineupConfirmedEvent.kt
@@ -1,4 +1,4 @@
-package com.nextup.core.event
+package com.nextup.core.domain.event
 
 /**
  * 라인업 확정 이벤트
