+++ nextup-core/src/main/kotlin/com/nextup/core/domain/event/TeamJoinApprovedEvent.kt
@@ -1,4 +1,4 @@
-package com.nextup.core.event
+package com.nextup.core.domain.event
 
 /**
  * 팀 가입 승인 이벤트
