+++ nextup-infrastructure/src/main/kotlin/com/nextup/infrastructure/service/attendance/AttendanceServiceImpl.kt
@@ -9,7 +9,7 @@ import com.nextup.core.domain.attendance.AttendancePoll
 import com.nextup.core.domain.attendance.AttendanceVote
 import com.nextup.core.domain.attendance.PollStatus
 import com.nextup.core.domain.attendance.VoteType
-import com.nextup.core.event.AttendanceVoteCreatedEvent
+import com.nextup.core.domain.event.AttendanceVoteCreatedEvent
 import com.nextup.core.port.attendance.AttendancePollRepositoryPort
 import com.nextup.core.port.attendance.AttendanceVoteRepositoryPort
 import com.nextup.core.port.repository.PlayerRepositoryPort
